move layout to utils
